Cleaner function export, better doc, "fix" macos/ios by disabling them
