+++ src/windows.rs
@@ -11,14 +11,17 @@ use std::ffi::CStr;
 use std::time::Duration;
 use std::{io, ptr};
 use windows_sys::Win32::Foundation::{
-    ERROR_BUFFER_OVERFLOW, ERROR_SUCCESS, HANDLE, NO_ERROR, WAIT_ABANDONED, WAIT_OBJECT_0, WAIT_TIMEOUT
+    ERROR_BUFFER_OVERFLOW, ERROR_SUCCESS, HANDLE, NO_ERROR, WAIT_ABANDONED, WAIT_OBJECT_0,
+    WAIT_TIMEOUT,
 };
 use windows_sys::Win32::NetworkManagement::IpHelper::{
     CancelIPChangeNotify, GetAdaptersAddresses, NotifyAddrChange, GAA_FLAG_INCLUDE_PREFIX,
     GAA_FLAG_SKIP_ANYCAST, GAA_FLAG_SKIP_DNS_SERVER, GAA_FLAG_SKIP_MULTICAST,
     IP_ADAPTER_ADDRESSES_LH, IP_ADAPTER_PREFIX_XP, IP_ADAPTER_UNICAST_ADDRESS_LH,
 };
-use windows_sys::Win32::Networking::WinSock::{WSACloseEvent, WSACreateEvent, WSAGetLastError, WSA_IO_PENDING};
+use windows_sys::Win32::Networking::WinSock::{
+    WSACloseEvent, WSACreateEvent, WSAGetLastError, WSA_IO_PENDING,
+};
 use windows_sys::Win32::System::Memory::{
     GetProcessHeap, HeapAlloc, HeapFree, HEAP_NONE, HEAP_ZERO_MEMORY,
 };
@@ -213,9 +216,6 @@ impl<'a> Iterator for UnicastAddressesIterator<'a> {
     }
 }
 
-/// Block until the OS reports that the network interface list has changed, or
-/// until an optional timeout. Returns an [`io::ErrorKind::WouldBlock`] error on
-/// timeout, or another error if the network notifier could not be set up.
 pub fn detect_interface_changes(timeout: Option<Duration>) -> io::Result<()> {
     let mut overlap: OVERLAPPED = unsafe { std::mem::zeroed() };
     let mut notify_event: HANDLE = Default::default();
@@ -239,7 +239,9 @@ pub fn detect_interface_changes(timeout: Option<Duration>) -> io::Result<()> {
 
     let ret = match unsafe { WaitForSingleObject(overlap.hEvent, millis) } {
         WAIT_OBJECT_0 => Ok(()),
-        WAIT_TIMEOUT | WAIT_ABANDONED => Err(io::Error::new(io::ErrorKind::WouldBlock, "Timed out")),
+        WAIT_TIMEOUT | WAIT_ABANDONED => {
+            Err(io::Error::new(io::ErrorKind::WouldBlock, "Timed out"))
+        }
         _ => Err(io::Error::last_os_error()),
     };
     unsafe {
